Refactored serializers folder.
