+++ api/views/home_views.py
@@ -5,7 +5,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
-from ..serializers import BookSerializer
+from ..serializers.model_serializers import BookSerializer
 from app.models import TheUser, AddedBook, Book
 from app.recommend import get_recommend
