+++ api/views/library_views.py
@@ -8,7 +8,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
-from ..serializers import CategorySerializer, BookSerializer
+from ..serializers.model_serializers import CategorySerializer, BookSerializer
 from app.models import TheUser, Category, Book
 
 OUTPUT_BOOKS_PER_PAGE = 20
