+++ api/views/profile_views.py
@@ -10,7 +10,7 @@ from rest_framework.decorators import api_view, parser_classes
 from rest_framework.parsers import MultiPartParser
 from rest_framework.response import Response
 
-from ..serializers import ProfileSerializer
+from ..serializers.model_serializers import ProfileSerializer
 from app.models import TheUser
 from app.tasks import changed_password
 from app.utils import resize_image
